+++ src/OpenTelemetry/Metrics/MetricPoint/HistogramBuckets.cs
@@ -63,9 +63,6 @@ public class HistogramBuckets
         this.BucketCounts = explicitBounds != null ? new HistogramBucketValues[explicitBounds.Length + 1] : Array.Empty<HistogramBucketValues>();
     }
 
-    private static double[]? CleanUpInfinitiesFromExplicitBounds(double[]? explicitBounds) => explicitBounds
-        ?.Where(b => !double.IsNegativeInfinity(b) && !double.IsPositiveInfinity(b)).ToArray();
-
     /// <summary>
     /// Returns an enumerator that iterates through the <see cref="HistogramBuckets"/>.
     /// </summary>
@@ -161,6 +158,9 @@ public class HistogramBuckets
             }
         }
     }
+
+    private static double[]? CleanUpInfinitiesFromExplicitBounds(double[]? explicitBounds) => explicitBounds
+        ?.Where(b => !double.IsNegativeInfinity(b) && !double.IsPositiveInfinity(b)).ToArray();
 
     /// <summary>
     /// Enumerates the elements of a <see cref="HistogramBuckets"/>.
